requirements.txt
matplotlib
scipy
scipy
numpy
opencv-python
ollama
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu124
huggingface-hub
transformers
tqdm
Pillow
accelerate
bitsandbytes
git+https://github.com/facebookresearch/sam2.git
decord
pandas
nltk
pycocoevalcap
seaborn
hf_xet
evaluate
rouge_score
bert-score
git+https://github.com/openai/CLIP.git
ipywidgets
tiktoken
blobfile
protobuf
sentencepiece
timm
num2words
flash-attn --no-build-isolation
